requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2
django-cities==0.6.1
django-model-utils==4.4.0
pillow==10.2.0
psycopg==3.1.18
psycopg-binary==3.1.18
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
Pillow